+++ lib/rdf/star/statement.ex
@@ -20,8 +20,6 @@ defmodule RDF.Star.Statement do
 
   @type t :: Triple.t() | Quad.t()
   @type coercible :: Triple.coercible() | Quad.coercible()
-  # deprecated: This will be removed in v0.11.
-  @type coercible_t :: coercible
 
   @type term_mapping :: RDF.Statement.term_mapping()